fix(ci): 修复 Install dependencies/ruff/mypy/pytest 四类 CI 失败
1. 打包：pyproject 显式 packages.find（api*/core*/parsers*），
   解决 flat-layout 多顶层包（含 frontend/）导致 editable 构建失败
2. ruff：test/ 排除出 lint 门禁 + per-file-ignores（N815/N803/B008）
   + 源码清理，ruff check/format 全绿
3. mypy：core/api/parsers 全部 65 处类型错误清零（54 文件 Success），
   含 ProviderInfo.client_class 改 Callable[..., AIClient]、
   策略列表基类标注、createdAt 时间戳→datetime 等
4. pytest：conftest 关闭测试进程限流（RATE_LIMIT_ENABLED=false），
   消除 RateLimitMiddleware 跨文件累计 60 次触发 429 的连锁失败；
   全量 652 passed + prod 认证回归 12 passed

+++ parsers/archive_parser.py
@@ -279,7 +279,7 @@ class ArchiveParser(BaseParser):
         ext = Path(filename).suffix.lower()
         return ext in text_exts
 
-    def _format_size(self, size: int) -> str:
+    def _format_size(self, size: float) -> str:
         """格式化文件大小"""
         for unit in ["B", "KB", "MB", "GB"]:
             if size < 1024:
